--- src/components/BottomSheetModal/index.js
@@ -1,4 +1,0 @@
-export { default as NaturalDisasterBottomSheet } from './NaturalDisasterBottomSheet';
-export { default as SocialDisasterBottomSheet } from './SocialDisasterBottomSheet';
-export { default as SelectRegionBottomSheet } from './SelectRegionBottomSheet';
-export { default as NotiContentBottomSheet } from './NotiContentBottomSheet';
