+++ src/navigation/TabNavigator.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
-import { Solution, Main, ReportPost } from '../pages';
+import { Solution, Main, ReportPost } from '../screens';
 import { RootTabParamList } from './types';
 
 const Tab = createBottomTabNavigator<RootTabParamList>();
