+++ src/screens/ReportArticleDetail.tsx
